feat: edition files connected

+++ src/Controller/FilesController.php
@@ -16,11 +16,13 @@ class FilesController extends AppController
      *
      * @return \Cake\Http\Response|null|void Renders view
      */
-    public function index()
+    public function index($editionID = null)
     {
-        $files = $this->paginate($this->Files);
+        if($editionID == null) return $this->redirect(['controller' => 'editions', 'action' => 'index']);
+        $files = $this->paginate($this->Files->findByEditionid($editionID));
 
         $this->set(compact('files'));
+        $this->set(compact('editionID'));
     }
 
     /**
@@ -30,13 +32,15 @@ class FilesController extends AppController
      * @return \Cake\Http\Response|null|void Renders view
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function view($id = null)
+    public function view($id = null, $editionID = null)
     {
+        if($editionID == null) return $this->redirect(['controller' => 'editions', 'action' => 'index']);
         $file = $this->Files->get($id, [
             'contain' => [],
         ]);
 
         $this->set(compact('file'));
+        $this->set(compact('editionID'));
     }
 
     /**
@@ -44,8 +48,9 @@ class FilesController extends AppController
      *
      * @return \Cake\Http\Response|null|void Redirects on successful add, renders view otherwise.
      */
-    public function add()
+    public function add($editionID = null)
     {
+        if($editionID == null) return $this->redirect(['controller' => 'editions', 'action' => 'index']);
         $file = $this->Files->newEmptyEntity();
         if ($this->request->is('post')) {
             $file = $this->Files->patchEntity($file, $this->request->getData());
@@ -62,16 +67,17 @@ class FilesController extends AppController
                 $file->name = $name;
             }
 
-            $file->editionId = 1;
+            $file->editionId = $editionID;
 
             if ($this->Files->save($file)) {
                 $this->Flash->success(__('The file has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                return $this->redirect(['action' => 'index', $editionID]);
             }
             $this->Flash->error(__('The file could not be saved. Please, try again.'));
         }
         $this->set(compact('file'));
+        $this->set(compact('editionID'));
     }
 
     /**
@@ -81,21 +87,38 @@ class FilesController extends AppController
      * @return \Cake\Http\Response|null|void Redirects on successful edit, renders view otherwise.
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function edit($id = null)
+    public function edit($id = null, $editionID = null)
     {
+        if($editionID == null) return $this->redirect(['controller' => 'editions', 'action' => 'index']);
         $file = $this->Files->get($id, [
             'contain' => [],
         ]);
         if ($this->request->is(['patch', 'post', 'put'])) {
             $file = $this->Files->patchEntity($file, $this->request->getData());
+            
+            if(!$file->getErrors){
+                $attachment = $this->request->getData('path');
+    
+                $name = $attachment->getClientFilename();
+    
+                $targetPath = WWW_ROOT.'img'.DS.$name;
+    
+                if($name) $attachment->moveTo($targetPath);
+    
+                $file->name = $name;
+            }
+           
+            $file->editionId = $editionID;
+
             if ($this->Files->save($file)) {
                 $this->Flash->success(__('The file has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                return $this->redirect(['action' => 'index', $editionID]);
             }
             $this->Flash->error(__('The file could not be saved. Please, try again.'));
         }
         $this->set(compact('file'));
+        $this->set(compact('editionID'));
     }
 
     /**
@@ -105,8 +128,9 @@ class FilesController extends AppController
      * @return \Cake\Http\Response|null|void Redirects to index.
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function delete($id = null)
+    public function delete($id = null, $editionID = null)
     {
+        if($editionID == null) return $this->redirect(['controller' => 'editions', 'action' => 'index']);
         $this->request->allowMethod(['post', 'delete']);
         $file = $this->Files->get($id);
         if ($this->Files->delete($file)) {
@@ -115,6 +139,6 @@ class FilesController extends AppController
             $this->Flash->error(__('The file could not be deleted. Please, try again.'));
         }
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect(['action' => 'index', $editionID]);
     }
 }
